Fix issue with scroll
When you are on a long page and scroll down, navigating to a new page
that is long enough, the new page's scroll position is also lower on the
page. That's a problem because important information is at the top of
the page. Navigating back must also not interfere with the stored scroll
position.

Based on a ChatGPT discussion at https://chatgpt.com/share/67b3c221-79d8-8001-9489-19e6aaed3012